Added resyncer.ts
